Removes standalone and AMD modularization in favor of a pure NPM module
