+++ .idea/jsLibraryMappings.xml
@@ -1,7 +1,3 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="JavaScriptLibraryMappings">
-    <file url="PROJECT" libraries="{Jasmine adapter for JsTestDriver}" />
-  </component>
-</project>
+<project version="4" />
