+++ package.json
@@ -1,24 +1,26 @@
 {
     "name": "pluto",
-    "description": "JavaScript dependency injection",
-    "homepage": "https://github.com/evan38109/Pluto.js",
+    "description": "Simple JavaScript dependency injection",
+    "homepage": "https://github.com/ecowden/pluto.js",
     "keywords": ["dependency", "injection", "di", "dependency injection", "testing"],
     "author": "Evan Cowden <evan38109@gmail.com>",
     "repository": {
         "type": "git",
-        "url": "git://github.com/evan38109/Pluto.js.git"
+        "url": "git://github.com/ecowden/pluto.js.git"
     },
-    "main": "src/pluto.js",
-    "version": "0.1.0",
+    "main": "lib/pluto.js",
+    "version": "0.2.0",
+    "engines": {
+        "node": "*"
+    },
     "dependencies": {
-        "underscore": "1.3.3"
+        "underscore": "1.5.2"
     },
     "devDependencies": {
-        "jasmine-node": "1.0.26",
-        "uglify": "1.3.2"
+        "jasmine-node": "1.11.0"
     },
     "scripts": {
-        "test": "jasmine-node test/spec",
-        "minify": "uglifyjs -o pluto.min.js src/pluto.js"
+        "test": "jasmine-node test",
+        "autotest": "jasmine-node --autotest test"
     }
 }
