--- test/testStandaloneScript.html
@@ -1,62 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>Jasmine Test Runner - Vanilla Script</title>
-
-    <!-- Jasmine -->
-    <link rel="stylesheet" type="text/css" href="jasmine/jasmine.css"/>
-    <link rel="icon" type="image/png" href="jasmine/jasmine_favicon.png">
-    <script type="text/javascript" src="jasmine/jasmine.js"></script>
-    <script type="text/javascript" src="jasmine/jasmine-html.js"></script>
-    <script type="text/javascript" src="jasmine/jasmine.teamcity_reporter.js"></script>
-
-    <!-- Third-party libraries -->
-    <script type="text/javascript" src="lib/underscore.js"></script>
-
-    <!-- Code under test -->
-    <script type="text/javascript" src="../src/pluto.js"></script>
-
-    <!-- Standalone test shim so that we don't need to re-write our specs -->
-    <script type="text/javascript" src="lib/define-standalone-test-shim.js"></script>
-
-    <!-- Specifications -->
-    <script type="text/javascript" src="spec/pluto.spec.js"></script>
-    <script type="text/javascript" src="spec/examples.spec.js"></script>
-
-</head>
-
-<body>
-<script type="text/javascript">
-    var jasmineEnv = jasmine.getEnv();
-    jasmineEnv.updateInterval = 1000;
-
-    var htmlReporter = new jasmine.HtmlReporter();
-    jasmine.getEnv().addReporter(htmlReporter);
-    jasmine.getEnv().addReporter(new jasmine.TeamcityReporter());
-
-    var OnCompleteReporter = _.extend(function () {}, jasmine.Reporter);
-    OnCompleteReporter.prototype.reportRunnerResults = function () {
-        console.log("##jasmine.reportRunnerResults");
-    };
-    jasmineEnv.addReporter(new OnCompleteReporter());
-
-    jasmineEnv.specFilter = function (spec) {
-        return htmlReporter.specFilter(spec);
-    };
-
-    var currentWindowOnload = window.onload;
-
-    window.onload = function () {
-        if (currentWindowOnload) {
-            currentWindowOnload();
-        }
-        execJasmine();
-    };
-
-    function execJasmine() {
-        jasmine.getEnv().execute();
-    }
-</script>
-
-</body>
-</html>
